Refactor OptiGlue 1.19.3
