+++ OptiGlue/1.19.3/src/main/kotlin/opekope2/optiglue_1_19_3/RegistryLookupServiceImpl.kt
@@ -1,4 +1,4 @@
-package opekope2.optiglue.mc_1_19_3
+package opekope2.optiglue_1_19_3
 
 import net.minecraft.block.Block
 import net.minecraft.entity.Entity
